# docs/openapi/openapi.yaml
openapi: 3.0.0
paths:
...
  /v1/auth/password/reset/request:
    post:
      description: Enqueues a password reset email for an existing user. Returns 204
        even if the email is unknown to avoid account enumeration.
      operationId: auth.password.reset.request
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PasswordResetRequestDto"
      responses:
        "204":
          description: ""
      summary: Request password reset
      tags: *a1
      x-error-codes:
        - VALIDATION_FAILED
        - RATE_LIMITED
        - INTERNAL
...
  /v1/me/profile-image/upload:
    post:
      description: Creates an upload record and returns a short-lived presigned URL
        for direct upload to object storage.
      operationId: users.me.profileImage.upload
      parameters:
        - name: Idempotency-Key
          in: header
          description: "Idempotency key for safe retries of write requests. Replays return
            `Idempotency-Replayed: true`."
          required: false
          schema:
            type: string
            minLength: 1
            maxLength: 128
            description: "Opaque string (recommended: UUIDv4)."
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProfileImageUploadRequestDto"
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProfileImageUploadPlanEnvelopeDto"
      security:
        - access-token: []
      summary: Create a profile image upload plan (presigned URL)
      tags: &a5
        - Users
      x-error-codes:
        - VALIDATION_FAILED
        - UNAUTHORIZED
        - RATE_LIMITED
        - IDEMPOTENCY_IN_PROGRESS
        - CONFLICT
        - USERS_OBJECT_STORAGE_NOT_CONFIGURED
        - INTERNAL
  /v1/me/profile-image/complete:
    post:
      description: Verifies the uploaded object and attaches it to the current user profile.
      operationId: users.me.profileImage.complete
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CompleteProfileImageUploadRequestDto"
      responses:
        "204":
          description: ""
      security:
        - access-token: []
      summary: Finalize a profile image upload
      tags: *a5
      x-error-codes:
        - VALIDATION_FAILED
        - UNAUTHORIZED
        - NOT_FOUND
        - USERS_OBJECT_STORAGE_NOT_CONFIGURED
        - USERS_PROFILE_IMAGE_NOT_UPLOADED
        - USERS_PROFILE_IMAGE_SIZE_MISMATCH
        - USERS_PROFILE_IMAGE_CONTENT_TYPE_MISMATCH
        - INTERNAL
  /v1/me/profile-image:
    delete:
      description: Detaches the current profile image (if any) and marks the stored
        file deleted. Object storage deletion is best-effort.
      operationId: users.me.profileImage.clear
      parameters: []
      responses:
        "204":
          description: ""
      security:
        - access-token: []
      summary: Clear current profile image
      tags: *a5
      x-error-codes:
        - UNAUTHORIZED
        - INTERNAL
  /v1/me/profile-image/url:
    get:
      description: Returns a short-lived presigned URL for rendering the current
        profile image. Returns 204 when no profile image is set.
      operationId: users.me.profileImage.url
      parameters: []
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProfileImageUrlEnvelopeDto"
        "204":
          description: ""
      security:
        - access-token: []
      summary: Get current profile image URL
      tags: *a5
      x-error-codes:
        - UNAUTHORIZED
        - USERS_OBJECT_STORAGE_NOT_CONFIGURED
        - INTERNAL
  /v1/me/account-deletion/request:
    post:
      description: Schedules account deletion 30 days in the future. The account
        remains usable during the grace period and the request can be canceled.
        After the grace period, the account is de-identified (PII erased) and
        the email becomes reusable.
      operationId: users.me.accountDeletion.request
      parameters:
        - name: Idempotency-Key
          in: header
          description: "Idempotency key for safe retries of write requests. Replays return
            `Idempotency-Replayed: true`."
          required: false
          schema:
            type: string
            minLength: 1
            maxLength: 128
            description: "Opaque string (recommended: UUIDv4)."
      responses:
        "204":
          description: ""
      security:
        - access-token: []
      summary: Request account deletion (30-day grace)
      tags: &a6
        - Users
      x-error-codes:
        - UNAUTHORIZED
        - IDEMPOTENCY_IN_PROGRESS
        - CONFLICT
        - USERS_CANNOT_DELETE_LAST_ADMIN
        - INTERNAL
...
info:
  title: Backend Core Kit API
  description: Generated API contract (code-first).
  version: 0.1.0
  contact: {}
...
servers:
  - url: /
components:
  securitySchemes:
    access-token:
      scheme: bearer
      bearerFormat: JWT
      type: http
      description: "First-party access token (Authorization: Bearer <token>)"
  schemas:
...
    PasswordResetRequestDto:
      type: object
      properties:
        email:
          type: string
          example: user@example.com
      required:
        - email
...
    CreateProfileImageUploadRequestDto:
      type: object
      properties:
        contentType:
          type: string
          enum:
            - image/jpeg
            - image/png
            - image/webp
          example: image/webp
          description: Declared MIME type for the upload. Must match the Content-Type used
            during upload.
        sizeBytes:
          type: number
          example: 123456
          minimum: 1
          maximum: 5000000
          description: Declared size of the upload in bytes (used to verify the stored
            object).
      required:
        - contentType
        - sizeBytes
    PresignedUploadDto:
      type: object
      properties:
        method:
          type: string
          enum:
            - PUT
          example: PUT
        url:
          type: string
          example: https://<r2-presigned-url>
        headers:
          type: object
          example:
            Content-Type: image/webp
          description: Headers that must be sent with the upload request.
      required:
        - method
        - url
        - headers
    ProfileImageUploadPlanDto:
      type: object
      properties:
        fileId:
          type: string
          example: 3d2c7b2a-2dd6-46a5-8f8e-3b5de8a5b0f0
        upload:
          $ref: "#/components/schemas/PresignedUploadDto"
        expiresAt:
          type: string
          example: 2026-01-10T12:34:56.789Z
          format: date-time
      required:
        - fileId
        - upload
        - expiresAt
    ProfileImageUploadPlanEnvelopeDto:
      type: object
      properties:
        data:
          $ref: "#/components/schemas/ProfileImageUploadPlanDto"
      required:
        - data
    CompleteProfileImageUploadRequestDto:
      type: object
      properties:
        fileId:
          type: string
          example: 3d2c7b2a-2dd6-46a5-8f8e-3b5de8a5b0f0
      required:
        - fileId
    ProfileImageUrlDto:
      type: object
      properties:
        url:
          type: string
          example: https://<r2-presigned-url>
        expiresAt:
          type: string
          example: 2026-01-10T12:34:56.789Z
          format: date-time
      required:
        - url
        - expiresAt
    ProfileImageUrlEnvelopeDto:
      type: object
      properties:
        data:
          $ref: "#/components/schemas/ProfileImageUrlDto"
      required:
        - data
...